basic doko element
